Refining project structure and file names
Projects are nice and easy to start working on again after
you get away of them after sometime, also it's easier to
others to understand how the project works and it'll be
easy to improve and maintain it over time. Also naming is
very important as well, they clarify instantly when someone
looks at it if it has a good straightforward name, otherwise
it'll get confusing and hard to figure out to change or edit.
KISS it so you're good to go once and forever.